requirements.txt
tweepy==4.14.0
yfinance==0.2.18
schedule==1.2.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.24.3 